CHATBOTV2/chatbot-generator/generator-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
requests
beautifulsoup4
python-multipart
